refactor(core/gpu): make GPU device a core-owned singleton and decouple from main

+++ engine/include/sustain/renderer/color.h
@@ -1,8 +1,10 @@
 #ifndef SUSTAIN_RENDERER_COLOR_H
 #define SUSTAIN_RENDERER_COLOR_H
 
+#include <stdint.h>
+
 typedef struct SN_Color {
-  float r, g, b, a;
+  uint8_t r, g, b, a;
 } SN_Color;
 
 #endif // !SUSTAIN_RENDERER_COLOR_H
